requirements.txt
glfw==2.7.0
numpy==2.0.1
PyGLM==2.7.1
PyOpenGL==3.1.7
pyrr==0.10.3
